+++ x/staking/handler.go
@@ -59,8 +59,8 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) ([]abci.ValidatorUpdate, sdk.T
 			continue
 		}
 
-		resTags.AppendTags(sdk.NewTags(
-			tags.Action, ActionCompleteUnbonding,
+		resTags = resTags.AppendTags(sdk.NewTags(
+			tags.Action, tags.ActionCompleteUnbonding,
 			tags.Delegator, dvPair.DelegatorAddress.String(),
 			tags.SrcValidator, dvPair.ValidatorAddress.String(),
 		))
@@ -75,7 +75,7 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) ([]abci.ValidatorUpdate, sdk.T
 			continue
 		}
 
-		resTags.AppendTags(sdk.NewTags(
+		resTags = resTags.AppendTags(sdk.NewTags(
 			tags.Action, tags.ActionCompleteRedelegation,
 			tags.Delegator, dvvTriplet.DelegatorAddress.String(),
 			tags.SrcValidator, dvvTriplet.ValidatorSrcAddress.String(),